finisheddddd

+++ src/Race.java
@@ -1,10 +1,11 @@
 /**
- * Race??
+ * Race
  */
 public class Race {
     private TwoThreeTreeRunner<RunnerID> IDTree;
     private TwoThreeTreeRunner<MinRunnerID> minTree;
     private TwoThreeTreeRunner<AvgRunnerID> avgTree;
+    private int count;
 
     public Race(){
         init();
@@ -20,6 +21,7 @@ public class Race {
         IDTree = new TwoThreeTreeRunner<>();
         minTree = new TwoThreeTreeRunner<>();
         avgTree = new TwoThreeTreeRunner<>();
+        count = 0;
     }
 
     /**
@@ -31,7 +33,10 @@ public class Race {
     public void addRunner(RunnerID id) // O(log(n))
     {
         if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("id cannot be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) != null)
+            throw new java.lang.IllegalArgumentException("runner already in race");
 
         /* create new Runner object and add the runner to each tree */
         Runner r = new Runner(id);
@@ -48,6 +53,11 @@ public class Race {
         IDTree.insert(node);
         minTree.insert(nodeMin);
         avgTree.insert(nodeAvg);
+
+        minTree.setMin_avg(minTree.minimum().getRunner());
+        avgTree.setMin_avg(avgTree.minimum().getRunner());
+
+        count++;
     }
 
     /**
@@ -59,7 +69,10 @@ public class Race {
     public void removeRunner(RunnerID id) // O(log(n))
     {
         if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("runner id can not be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and remove them from the min and avg trees */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
@@ -70,6 +83,9 @@ public class Race {
         avgTree.delete(r.getAvgLeaf());
         r.setAvgLeaf(null);
         IDTree.delete(node);
+        minTree.setMin_avg(minTree.minimum().getRunner());
+        avgTree.setMin_avg(avgTree.minimum().getRunner());
+        count--;
     }
 
     /**
@@ -81,12 +97,18 @@ public class Race {
      */
     public void addRunToRunner(RunnerID id, float time) // O(log(n) + log(m))
     {
-        if(id == null || time <= 0)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+        if(id == null || time < 0)
+            throw new java.lang.IllegalArgumentException("id cant be null, time cant be negative");
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("no runner of that ID in race");
 
         /* find the runner in the IDs tree and add a run to their Runner object */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
         Runner r = node.getRunner();
+
+        if(r.runExists(time))
+            throw new java.lang.IllegalArgumentException("run of that time already exists");
+
         r.addRun(time);
         //update the Trees
         this.updateTrees(r);
@@ -104,14 +126,22 @@ public class Race {
     public void removeRunFromRunner(RunnerID id, float time) // O(log(n) + log(m))
     {
         if (id == null || time <= 0)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("not implemented");
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and remove the run */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
         Runner r = node.getRunner();
+
+        if(!r.runExists(time))
+            throw new java.lang.IllegalArgumentException("can't remove run that does not exist");
+
         r.deleteRun(time);
         //update the Trees
         this.updateTrees(r);
+        minTree.setMin_avg(minTree.minimum().getRunner());
+        avgTree.setMin_avg(avgTree.minimum().getRunner());
     }
 
     /**
@@ -122,6 +152,9 @@ public class Race {
      */
     public RunnerID getFastestRunnerAvg() // O(1)
     {
+        if(count==0)
+            throw new java.lang.IllegalArgumentException("no runners in race");
+
         /* lowest runtime average = key of the root in the avg tree */
         return avgTree.getMin_avg();
     }
@@ -134,6 +167,9 @@ public class Race {
      */
     public RunnerID getFastestRunnerMin() // O(1)
     {
+        if(count==0)
+            throw new java.lang.IllegalArgumentException("no runners in race");
+
         /* lowest minimal runtime = key of the root in the min tree */
         return minTree.getMin_avg();
     }
@@ -148,7 +184,10 @@ public class Race {
     public float getMinRun(RunnerID id) // O(log(n))
     {
         if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("id can't be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and get their minimal runtime */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
@@ -166,7 +205,10 @@ public class Race {
     public float getAvgRun(RunnerID id) // O(log(n))
     {
         if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("runner id can't be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and get their avg runtime */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
@@ -183,13 +225,16 @@ public class Race {
      */
     public int getRankAvg(RunnerID id) // O(log(n))
     {
-        if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+        if(id == null)
+            throw new java.lang.IllegalArgumentException("runner id cannot be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and find their rank in the avg tree */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
         Runner r = node.getRunner();
-        return avgTree.rank(r.getAvgLeaf());
+        return count - avgTree.rank(r.getAvgLeaf()) + 1;
     }
 
     /**
@@ -202,12 +247,15 @@ public class Race {
     public int getRankMin(RunnerID id) // O(log(n))
     {
         if (id == null)
-            throw new java.lang.UnsupportedOperationException("not implemented");
+            throw new java.lang.IllegalArgumentException("id cannot be null");
+
+        if (IDTree.search(IDTree.getRoot(), id) == null)
+            throw new java.lang.IllegalArgumentException("No runner of that id in race");
 
         /* find the runner in the IDs tree and find their rank in the min tree */
         NodeRunner<RunnerID> node = (NodeRunner<RunnerID>)IDTree.search(IDTree.getRoot(), id);
         Runner r = node.getRunner();
-        return minTree.rank(r.getMinLeaf());
+        return count-minTree.rank(r.getMinLeaf()) + 1;
     }
 
     /**
